+++ src/common/logger.go
@@ -6,9 +6,8 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// This can be used as the destination for a logger and it'll
-// map them into calls to testing.T.Log, so that you only see
-// the logging for failed tests.
+// This can be used as the destination for a logger and it'll map them into
+// calls to testing.T.Log, so that you only see the logging for failed tests.
 type testLoggerAdapter struct {
 	t      *testing.T
 	prefix string