+++ src/state/state.go
@@ -47,7 +47,7 @@ type State struct {
 	logger *logrus.Logger
 }
 
-func NewState(logger *logrus.Logger, dbFile string, dbCache int, genesisFile string) (*State, error) {
+func NewState(dbFile string, dbCache int, genesisFile string, logger *logrus.Logger) (*State, error) {
 	handles, err := getFdLimit()
 	if err != nil {
 		return nil, err
@@ -260,12 +260,10 @@ func (s *State) GetNonce(addr common.Address) uint64 {
 	return s.ethState.GetNonce(addr)
 }
 
-
 // GetCode returns an account's bytecode from the main ethState
 func (s *State) GetCode(addr common.Address) []byte {
 	return s.ethState.GetCode(addr)
 }
-
 
 // GetPoolNonce returns an account's nonce from the txpool's ethState
 func (s *State) GetPoolNonce(addr common.Address) uint64 {